+++ examples/java/cdap/hubspot/src/main/java/org/apache/beam/examples/complete/cdap/hubspot/README.md
@@ -15,14 +15,6 @@
     under the License.
 -->
 
-# Apache Beam pipeline examples for CdapIO and CDAP plugins
-
-This directory contains set of [Apache Beam](https://beam.apache.org/) pipeline examples to read data
-from a [CDAP plugin](https://github.com/data-integrations) and write data into .txt file (and vice versa).
-
-Supported CDAP plugins:
-- [Hubspot](https://github.com/data-integrations/hubspot)
-
 ## Gradle preparation
 
 To run this example your `build.gradle` file should contain the following task to execute the pipeline:
@@ -41,7 +33,7 @@ task executeCdap (type:JavaExec) {
 Gradle 'executeCdap' task allows to run the pipeline via the following command:
 
 ```bash
-gradle clean executeCdap -DmainClass=org.apache.beam.examples.complete.cdap.CdapHubspotToTxt \
+gradle clean executeCdap -DmainClass=org.apache.beam.examples.complete.cdap.hubspot.CdapHubspotToTxt \
      -Dexec.args="--<argument>=<value> --<argument>=<value>"
 ```
 
@@ -66,7 +58,7 @@ To execute this pipeline, specify the parameters in the following format:
 Gradle 'executeCdap' task allows to run the pipeline via the following command:
 
 ```bash
-gradle clean executeCdap -DmainClass=org.apache.beam.examples.complete.cdap.CdapHubspotStreamingToTxt \
+gradle clean executeCdap -DmainClass=org.apache.beam.examples.complete.cdap.hubspot.CdapHubspotStreamingToTxt \
      -Dexec.args="--<argument>=<value> --<argument>=<value>"
 ```
 
@@ -95,7 +87,7 @@ To execute this pipeline, specify the parameters in the following format:
 Gradle 'executeCdap' task allows to run the pipeline via the following command:
 
 ```bash
-gradle clean executeCdap -DmainClass=org.apache.beam.examples.complete.cdap.TxtToCdapHubspot \
+gradle clean executeCdap -DmainClass=org.apache.beam.examples.complete.cdap.hubspot.TxtToCdapHubspot \
      -Dexec.args="--<argument>=<value> --<argument>=<value>"
 ```
