+++ examples/java/cdap/hubspot/src/main/java/org/apache/beam/examples/complete/cdap/hubspot/options/CdapHubspotOptions.java
@@ -15,30 +15,32 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.complete.cdap.options;
+package org.apache.beam.examples.complete.cdap.hubspot.options;
 
+import org.apache.beam.examples.complete.cdap.options.BaseCdapOptions;
 import org.apache.beam.sdk.options.Description;
 import org.apache.beam.sdk.options.Validation;
 
 /**
- * The {@link CdapHubspotSinkOptions} interface provides the custom execution options passed by the
- * executor at the command-line for {@link org.apache.beam.examples.complete.cdap.TxtToCdapHubspot}
- * example.
+ * The {@link CdapHubspotOptions} interface provides the custom execution options passed by the
+ * executor at the command-line for examples with Cdap Hubspot plugins.
  */
-public interface CdapHubspotSinkOptions extends CdapHubspotOptions {
+public interface CdapHubspotOptions extends BaseCdapOptions {
+
+  @Description("Hubspot api server url. If not specified then the default url will be used.")
+  String getApiServerUrl();
+
+  void setApiServerUrl(String apiServerUrl);
 
   @Validation.Required
-  @Description("Input .txt file path with Hubspot records.")
-  String getInputTxtFilePath();
+  @Description("Hubspot Private Application Access Token.")
+  String getAuthToken();
 
-  void setInputTxtFilePath(String inputTxtFilePath);
+  void setAuthToken(String authToken);
 
   @Validation.Required
-  @Description(
-      "Locks directory path where locks will be stored."
-          + "This parameter is needed for Hadoop External Synchronization"
-          + "(mechanism for acquiring locks related to the write job).")
-  String getLocksDirPath();
+  @Description("Name of object to pull from Hubspot (e.g. Contacts).")
+  String getObjectType();
 
-  void setLocksDirPath(String locksDirPath);
+  void setObjectType(String objectType);
 }
