+++ examples/java/cdap/hubspot/src/main/java/org/apache/beam/examples/complete/cdap/hubspot/package-info.java
@@ -15,24 +15,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.complete.cdap.options;
 
-import org.apache.beam.sdk.options.Description;
-
-/**
- * The {@link CdapHubspotSourceOptions} interface provides the custom execution options passed by
- * the executor at the command-line for {@link
- * org.apache.beam.examples.complete.cdap.CdapHubspotStreamingToTxt} example.
- */
-public interface CdapHubspotStreamingSourceOptions extends CdapHubspotSourceOptions {
-
-  @Description("Delay in seconds between polling for new records updates.")
-  Long getPullFrequencySec();
-
-  void setPullFrequencySec(Long pullFrequencySec);
-
-  @Description("Inclusive start offset from which the reading should be started.")
-  Long getStartOffset();
-
-  void setStartOffset(Long startOffset);
-}
+/** Cdap Plugins template. */
+package org.apache.beam.examples.complete.cdap.hubspot;
